refactor and organization
