+++ .idea/workspace.xml
@@ -6,7 +6,9 @@
   <component name="ChangeListManager">
     <list default="true" id="a7ac25b6-265d-43ed-8230-0c6749f7127e" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/java/Parser/Parser.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Parser/Parser.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/Parser/ParserClass.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Parser/ParserClass.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/Parser/ParserDriver.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Parser/ParserDriver.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/Parser/ParserUtilities.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/Parser/ParserUtilities.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -69,9 +71,9 @@
       <recent name="Parser" />
     </key>
   </component>
-  <component name="RunManager" selected="Application.ParserDriver">
-    <configuration name="Parser" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="Parser.Parser" />
+  <component name="RunManager" selected="Application.Parser.ParserDriver">
+    <configuration name="ParserClass" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="Parser.ParserClass" />
       <module name="MiniCompiler.main" />
       <extension name="coverage">
         <pattern>
@@ -89,6 +91,19 @@
       <extension name="coverage">
         <pattern>
           <option name="PATTERN" value="Parser.OOP_parser.*" />
+          <option name="ENABLED" value="true" />
+        </pattern>
+      </extension>
+      <method v="2">
+        <option name="Make" enabled="true" />
+      </method>
+    </configuration>
+    <configuration name="Parser.ParserDriver" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
+      <option name="MAIN_CLASS_NAME" value="Parser.ParserDriver" />
+      <module name="MiniCompiler.main" />
+      <extension name="coverage">
+        <pattern>
+          <option name="PATTERN" value="Parser.*" />
           <option name="ENABLED" value="true" />
         </pattern>
       </extension>
@@ -124,29 +139,13 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <configuration name="LexerJunitTests.simpleAssert" type="JUnit" factoryName="JUnit" temporary="true" nameIsGenerated="true">
-      <module name="MiniCompiler.test" />
-      <extension name="coverage">
-        <pattern>
-          <option name="PATTERN" value="Lexer.*" />
-          <option name="ENABLED" value="true" />
-        </pattern>
-      </extension>
-      <option name="PACKAGE_NAME" value="Lexer" />
-      <option name="MAIN_CLASS_NAME" value="Lexer.LexerJunitTests" />
-      <option name="METHOD_NAME" value="simpleAssert" />
-      <option name="TEST_OBJECT" value="method" />
-      <method v="2">
-        <option name="Make" enabled="true" />
-      </method>
-    </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="Application.Parser.ParserDriver" />
         <item itemvalue="Application.ParserDriver" />
-        <item itemvalue="Application.Parser" />
+        <item itemvalue="Application.ParserClass" />
         <item itemvalue="Application.ProgramRunner" />
         <item itemvalue="JUnit.LexerJunitTests" />
-        <item itemvalue="JUnit.LexerJunitTests.simpleAssert" />
       </list>
     </recent_temporary>
   </component>
@@ -158,7 +157,7 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1678159185811</updated>
-      <workItem from="1684133509663" duration="42225000" />
+      <workItem from="1684133509663" duration="44645000" />
     </task>
     <servers />
   </component>
