+++ src/main/resources/additional_input_files/nfactorial.c
