+++ src/main/resources/example_input_files/count.c
