+++ src/main/resources/example_input_files/hello.lex
