+++ src/main/resources/example_input_files/loop.py
