--- src/main/resources/hello_test.par
@@ -1,7 +1,0 @@
-Sequence
-;
-Sequence
-;
-Prts
-String "Hello, World!\n" 
-;
