--- src/main/resources/loop_test.par
@@ -1,31 +1,0 @@
-Sequence
-Sequence
-;
-Assign
-Identifier count 
-Integer 1 
-While
-Less
-Identifier count 
-Integer 10 
-Sequence
-Sequence
-;
-Sequence
-Sequence
-Sequence
-;
-Prts
-String "count is: " 
-;
-Prti
-Identifier count 
-;
-Prts
-String "\n" 
-;
-Assign
-Identifier count 
-Add
-Identifier count 
-Integer 1
