+++ src/main/resources/primary_input/prime.c
